Improve Selectors logic
Create the abstract class ObjectSelector that inherits from Selector. In
this class implemented the GetObject method, and now the children
classes just have to edit the IsSelectable method, that returns true or
false depending if the obtained object is the object the child class
wants to select yes or no.